# Improve performance for the binary decoder