The modified version is approximately 72% faster than the original, with an average time of 2.8935 ms compared to 10.5994 ms for the original. The total time taken by the modified version was also significantly less, at 5034.9741 ms compared to 31966.7486 ms for the original.

Here are the enhancements:

- Instead of concating the string and creating a new string on every byte iteration, we used a table to store every character without creating multiple new strings
- Instead of iterating the string using `string.gmatch` function, we are now using a numeric `for` loop for iterating every 8 characters (every byte) which is much faster than `string.gmatch`. Keep in mind that we are now using `string.sub` function to get the bytes of the given encoded string.
- Returning the concated version of the structured character table once finished